Fixed 'default' session handling.
Our environment is not anymore only 'session' aware but also has a 'project' context.
This 'project' context is given by the existence of '.yavide_proj' file which defines
project specific settings. Problem arises when 'default' session is being loaded.
As this session is basically not a real project, system will complain that it cannot
find '.yavide_proj'. Therefore, a dummy folder containing '.yavide_proj' associated
with 'default' session had to be created.

This is a restriction of 'vim-session', a plugin which has been utilized to implement project
awareness. In future, I would like to replace this plugin with my own implementation
and hence remove this restriction.