Reimplemented project workspace management.
Up until now 'vim-session' plugin has been in charge of handling project workspaces.
However, it didn't fit with all of its diverse features into this environment. For example,
it wasn't possible to save session files into the respective project directories, but only to
a common directory supplied by a user. Also, it has a weird restriction which requires the
environment to always have a 'default' session which is basically redundant as it does not
really fit into any project context.

Therefore, it has been decided to reimplement this functionality from the scratch. So, this commit
provides the initial implementation, which currently may not be as featureful as 'vim-session' but
it does its job. There are some minor things left to be implemented (i.e. locking the session,
deleting the project). Implementation is part of the 'Project' API.

Besides the new implementation which tries to resemble old use-cases, there are some other differences
which change 'Yavide' behavior. For example, 'SessionLoadPost' event is not anymore used to
trigger the automatic loading of last project. New behavior will enable user to decide which
recently opened project he wants to re-open. Also, 'default' session restriction has been removed
as well as need for the 'default' project. Other additional 'vim-session'-like user defined
options will be included as well in features to follow.

Consequence of these changes is also a much better 'Buffer' API implementation. Many things
had to be improved in order to provide more error-prone environment. Specifically,
many corner cases when closing and/or saving files have been fixed.