Introduced a fix for 'mixed' type projects.
Prior to this change, source indexer has not been properly initialized with programming languages
used throughout the project. This is now fixed by 'auto-detecting' the programming languages.
However, this is only a temporary fix and will be handled manually by a user as soon as new interface
for project management becomes stable.