Fixed compatibility issue.
Lack of 'servername' feature compiled into Vim will cause a bug in 'YavideUtils.send_vim_remote_command()'
as it directly utilizes this feature for cross-process communication. Therefore, every module depending
on this API will not function properly (i.e. source code indexer). This kind of failure has been
observed on Fedora machines. Ubuntu machines, for example, didn't exhibit this failure. However,
it seems that GUI versions of Vim such as 'gvim' have this feature compiled in by default.

To be on a safe side best would be to test for this feature during the installation process.