Improved installation process.
Sometimes it might happen that package manager aborts the installation of some package.
This will introduce an unresolved dependency for the 'Yavide' environment. As it looks
like this situation occurs only when there is a question [Y/n] if one really wants to
install the package. To circumvent this problem an additional flag has been introduced
which basically forces the package manager to automatically install the package
with no questions.