Implementation of generic client-server (async) framework. Adaption of syntax highlighting to the new framework.
Even though not polished to the fullest, this framework enables integration of any kind of services in
an *asynchronous* fashion into the environment. That means that any time- and/or resource-consuming
service will *not* be blocking a main thread UI but it will be rather dispatched to the separate
process.

How it is designed is conceptually fairly simple:
  * Service
    * Has its own unique ID (it must match both on the client (Yavide) and server (YavideServer) side).
    * Runs in its own background process.
    * Can be started-up and shut-down on demand.
    * Can be triggered during its runtime.
    * Can notify the main thread about its own events and queue the actions to be executed on the UI side.
  * 'Yavide'
    * Starts all the configured services during the environment start-up.
    * Executes actions, if any, queued by particular services and processes the results.
    * Shuts down all the configured services during the environment shut-down.

This is really a big step towards to the seamless integration of any IDE-like service such as:
  * Syntax highlighting (already done)
  * Indexing
  * Linting
  * Auto-formatting
  * Building
  * Fix-it's
  * On-the-fly compilation errors
  * Or any other action that should be executing asynchronously in the background (i.e. grep)