Configure 'ProjectBuilder' service with the build command when triggering the build.
Previously it was done during the service startup, which takes place upon project loading, but this
assumed that build command was already available and this is not the case when, for example,
importing the project or creating a new project. Service in that case has been already started with
the empty build command string (because it was not set) and would have to be re-configured
once the build command was set by the user to apply the new changes. I found the other way to be
more feasible as I don't find introducing a new interface really neccessary at this moment.