Fix a Clang parsing error triggered by serializing modified buffer contents into the temp file.
Now, when we settled down for using temp files in order to trigger analysis on modified files,
we have to make sure that all the header includes of that particular file are not invalidated
(e.g. not found). This happens because what we basically have done is that we moved a file outside
the project directory structure.

A fix is to include additional include path to compiler args which points to the parent directory
of file being modified.

Alternative would be to generate temp files in original location but that would pollute project
directory and potentially would not play well with some other tools (i.e. indexer, version control, etc.)