Integrate 'working-directory' compiler flag.
We set the working-directory to project root directory. This makes sense because
otherwise compiler might be unable to resolve relevant includes. I.e. when not
run from the root of project directory.

Ideally, this parameter, as well as compiler args, shall be passed during
'YavideService::startup_hook()' phase but due to 'SyntaxHighlighterService' being
run as soon as 'BufEnter' event is triggered there is a race condition. In other words,
'BufEnter' might be triggered before 'SyntaxHighlighterService' is initialized.
We will need some kind of a guard in order to circumvent these kind of issues.