Implement proper CursorKind.OVERLOADED_DECL_REF handling.
Cursors which are of CursorKind.OVERLOADED_DECL_REF don't get us any meaningful information
about the code without extra handling.

CursorKind.OVERLOADED_DECL_REF basically identifies a reference to a set of overloaded functions
or function templates which have not yet been resolved to a specific function or function template.
This means that cursor kind might be one of the following:
  CursorKind.FUNCTION_DECL, CursorKind.FUNCTION_TEMPLATE, CursorKind.CXX_METHOD
To extract more information about the cursor we have to use `clang_getNumOverloadedDecls()` to get how
many overloads there are and then use `clang_getOverloadedDecl()` to get a specific overload.

In our case, we can always use the first overload because it is expected (not possible?) that a set of
overloads cannot be made of semantically different kinds (i.e. CursorKind.FUNCTION_DECL, CursorKind.CXX_METHOD).

Additional note:
  * `libclang` Python bindings do not expose the implementation of `clang_getNumOverloadedDecls()` and `clang_getOverloadedDecl()`
  * Until there is a patch, these interfaces are only temporarily implemented as part of `ClangTokenizer` class