Implemented clang.cindex AST traversal interface. Monkey-patched existing clang.cindex.Cursor.get_children() interface.
New AST traversal interface, `traverse(client_data, client_visitor)`, is more flexible and resembles the one found
in libclang C API. It enables us to write custom visitor functions which allows us to do whatever we want with
given node (cursor). I.e. we might only want to append the node to the cursor list, just like we do in `get_children()`,
but we might also want to implement a different logic for handling the cursors and not only appending them to the list.

`clang.cindex.Cursor.get_children()` has been monkey-patched with `get_children_patched()`. This is required in order to
implement some features. More specifically, sometimes we will need an information about the parent of given AST node and
this is what this implementation brings to the table. Official implementation does not store such information about the
AST nodes.