Implement clang diagnostics/fixits service. Make it play with Vim QuickFix.
This is a second service now for which we are supposed to be providing feedback
as we are typing (the same is with semantic syntax highlighting).

Therefore, some refactoring had to take place in order to put these under a common
denominator: now there is a common, 'TextChangedI', event which we are now using to
hook all services on which are supposed to be running like this. This avoids code
duplication and makes it easier for future integrations.