Do not run indexer on existing (stored) indexer results.
In order to easily reconstruct the original filename when loading existing
indexing results, indexing results were being stored with the extension being
equal to the extension of a file that was operated on. However, this could
cause some subtle bugs so from now on '.ast' extension will be used to store
the indexing results.