Files to be indexed are now passed through the temp. file to the indexing subprocess.
Previous approach passed this information via raw string which caused troubles
if string is too long (exception generated). The maximum length of this string
that can be passed through is platform dependent and this limit is very easy
to hit when indexing bigger projects.

Alternative way to cope with this situation is to actually serialized a list of
files to be indexed in a temporary file and pass that temporary file through.