Multiple identical symbols on the same line will be stored only once.
Having primary key defined as (filename, usr, line) and not
(filename, usr, line, column) is advantageous because of several reasons.

We may get multiple instances of the same symbol (e.g. of the same USR)
as part of multiple expressions in the AST tree for that particular line
of code. These instances will only differentiate by the column they start
with but this is essentially not the information we are interested in.

To identify such 'redundant' symbols we would have to employ some kind of
heuristics in order to avoid such symbols. Easy way to fix this issue
is to say: same symbol can occur multiple times in the same line but we
will store only one of them.