Implement proper service startup/shutdown sequence.
Service startup/shutdown callbacks have been introduced. Startup/shutdown
hooks are not anymore optional but it is rather anticipated that every
service will want to have an init/deinit hook. This also enables us to
implement a proper init/deinit sequence on the client side (i.e. not
being able to trigger the service until it has been initialized).

Additionaly, shutdown callback will optionally reply to the master
as there are cases where it is not always required to invoke it (i.e.
when shutting down the whole application).