Automate the step further with compilation database setup.
We now automatically detect if there is any compilation db in the project
root directory and, if there is one, then we utilize it for source code model.

If there was none found, we issue a warning to user about limited functionality.

Also, if there was no compilation database found initially when project was
imported/created, but it was generated in the meantime, we will automatically
check that with the next run of Yavide.