Implement better compiler args handling for files not found in the JSON DB.
Whenever there is no corresponding JSON entry for the given file, just try
to use compiler args of the very first entry in the JSON DB. If that fails as
well (i.e. header-only libs), just use the default args that we can extract
from the system.

This approach is going to be valid only if different files are not going to be
compiled in different ways. I.e. with different flags, different include paths,
etc.

A following bug has been also fixed with this patch indirectly:
  (1) Open the project with the header file opened by default.
  (2) No entry will be found in the JSON and 'default args' will be used.
  (3) That will be cached in the translation unit cache.
  (4) For every next time this cached item will be re-used, with, obviously,
      incomplete flags even though we might have a more complete args by then.
      I.e. we might have opened a cpp file which would get us to cache its
      compiler args.